+++ openyich-data/src/main/java/com/openyich/framework/data/filter/Filter.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 import java.io.Serializable;
 import java.util.List;
