+++ openyich-data/src/main/java/com/openyich/framework/data/filter/FloatFilter.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 /**
  * Filter class for {@link Float} type attributes.
