+++ openyich-data/src/main/java/com/openyich/framework/data/filter/IntegerFilter.java
@@ -1,4 +1,4 @@
-package com.openyich.framework.data.jpa.filter;
+package com.openyich.framework.data.filter;
 
 /**
  * Filter class for {@link Integer} type attributes.
