+++ openyich-data/src/test/java/com/openyich/framework/data/jpa/filter/BigDecimalFilterTest.java
@@ -9,6 +9,9 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.openyich.framework.data.filter.BigDecimalFilter;
+import com.openyich.framework.data.filter.Filter;
+
 public class BigDecimalFilterTest {
 
   private BigDecimalFilter filter;
