+++ openyich-data/src/test/java/com/openyich/framework/data/jpa/filter/FilterTest.java
@@ -1,6 +1,8 @@
 package com.openyich.framework.data.jpa.filter;
 
 import com.google.common.collect.Lists;
+import com.openyich.framework.data.filter.Filter;
+
 import org.junit.Before;
 import org.junit.Test;
