+++ openyich-data/src/test/java/com/openyich/framework/data/jpa/filter/InstantFilterTest.java
@@ -9,6 +9,9 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.openyich.framework.data.filter.Filter;
+import com.openyich.framework.data.filter.InstantFilter;
+
 public class InstantFilterTest {
 
   private InstantFilter filter;
