+++ openyich-data/src/test/java/com/openyich/framework/data/jpa/filter/LocalDateFilterTest.java
@@ -9,6 +9,9 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.openyich.framework.data.filter.Filter;
+import com.openyich.framework.data.filter.LocalDateFilter;
+
 public class LocalDateFilterTest {
 
   private LocalDateFilter filter;
