+++ openyich-data/src/test/java/com/openyich/framework/data/jpa/filter/ShortFilterTest.java
@@ -8,6 +8,9 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.openyich.framework.data.filter.Filter;
+import com.openyich.framework.data.filter.ShortFilter;
+
 public class ShortFilterTest {
 
   private ShortFilter filter;
